refactor(website): split discovery route modules
