+++ website/app/routes/languages.$language.tsx
@@ -1,0 +1,2 @@
+export { loader } from '@/features/discovery/loader.server';
+export { default, links, meta } from '@/features/discovery/route';
